backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Marshmallow==0.15.0
Marshmallow-SQLAlchemy==0.29.0
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
marshmallow==3.20.1
Pillow==10.0.1
cryptography==41.0.4